rename: util 패키지를 세부패키지로 구분
