+++ src/main/java/vendingmachine/util/parser/Parser.java
@@ -1,4 +1,6 @@
-package vendingmachine.util;
+package vendingmachine.util.parser;
+
+import vendingmachine.util.validator.CoinValidator;
 
 public class Parser {
 	public static int convertStringToInt(String input) throws IllegalArgumentException {
